# docs/api/openapi.snapshot.json
{
  "components": {
    "schemas": {
      "AIDLPFinding": {
        "additionalProperties": false,
        "properties": {
          "confidence": {
            "title": "Confidence",
            "type": "number"
          },
          "detector": {
            "title": "Detector",
            "type": "string"
          },
          "finding_type": {
            "title": "Finding Type",
            "type": "string"
          },
          "mitre": {
            "items": {
              "type": "string"
            },
            "title": "Mitre",
            "type": "array"
          },
          "severity": {
            "title": "Severity",
            "type": "string"
          },
          "sha256_prefix": {
            "title": "Sha256 Prefix",
            "type": "string"
          }
        },
        "required": [
          "detector",
          "finding_type",
          "severity",
          "confidence",
          "mitre",
          "sha256_prefix"
        ],
        "title": "AIDLPFinding",
        "type": "object"
      },
      "AIPipelineMeta": {
        "additionalProperties": false,
        "properties": {
          "attack": {
            "title": "Attack",
            "type": "string"
          },
          "combined_detection": {
            "default": false,
            "title": "Combined Detection",
            "type": "boolean"
          },
          "dlp_finding_count": {
            "default": 0,
            "title": "Dlp Finding Count",
            "type": "integer"
          },
          "dlp_findings": {
            "items": {
              "$ref": "#/components/schemas/AIDLPFinding"
            },
            "title": "Dlp Findings",
            "type": "array"
          },
          "dlp_max_severity": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dlp Max Severity"
          },
          "llm_mode": {
            "title": "Llm Mode",
            "type": "string"
          },
          "mode": {
            "title": "Mode",
            "type": "string"
          },
          "risk_action": {
            "default": "allow",
            "title": "Risk Action",
            "type": "string"
          },
          "risk_score": {
            "default": 0,
            "title": "Risk Score",
            "type": "integer"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          }
        },
        "required": [
          "run_id",
          "attack",
          "mode",
          "llm_mode"
        ],
        "title": "AIPipelineMeta",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "IncidentEvidenceNew": {
        "properties": {
          "counts": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Counts",
            "type": "object"
          },
          "events": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Events",
            "type": "array"
          },
          "timeline": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Timeline",
            "type": "array"
          },
          "window_end": {
            "title": "Window End",
            "type": "string"
          },
          "window_start": {
            "title": "Window Start",
            "type": "string"
          }
        },
        "required": [
          "window_start",
          "window_end",
          "counts",
          "timeline",
          "events"
        ],
        "title": "IncidentEvidenceNew",
        "type": "object"
      },
      "IncidentExplanationNew": {
        "properties": {
          "observed": {
            "title": "Observed",
            "type": "integer"
          },
          "threshold": {
            "title": "Threshold",
            "type": "integer"
          },
          "trigger_field": {
            "title": "Trigger Field",
            "type": "string"
          },
          "window": {
            "title": "Window",
            "type": "string"
          }
        },
        "required": [
          "threshold",
          "observed",
          "window",
          "trigger_field"
        ],
        "title": "IncidentExplanationNew",
        "type": "object"
      },
      "IncidentInvestigationNew": {
        "properties": {
          "escalation_conditions": {
            "items": {
              "type": "string"
            },
            "title": "Escalation Conditions",
            "type": "array"
          },
          "false_positives": {
            "items": {
              "type": "string"
            },
            "title": "False Positives",
            "type": "array"
          },
          "pivot_queries": {
            "items": {
              "$ref": "#/components/schemas/InvestigationPivotQuery"
            },
            "title": "Pivot Queries",
            "type": "array"
          },
          "questions": {
            "items": {
              "type": "string"
            },
            "title": "Questions",
            "type": "array"
          }
        },
        "required": [
          "questions",
          "pivot_queries",
          "false_positives",
          "escalation_conditions"
        ],
        "title": "IncidentInvestigationNew",
        "type": "object"
      },
...
      "IncidentMitreMappingNew": {
        "properties": {
          "tactic": {
            "title": "Tactic",
            "type": "string"
          },
          "technique": {
            "title": "Technique",
            "type": "string"
          },
          "technique_name": {
            "title": "Technique Name",
            "type": "string"
          }
        },
        "required": [
          "tactic",
          "technique",
          "technique_name"
        ],
        "title": "IncidentMitreMappingNew",
        "type": "object"
      },
...
      "IncidentPatchRequest": {
        "additionalProperties": false,
        "properties": {
          "assignee": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assignee"
          },
          "reopen": {
            "default": false,
            "title": "Reopen",
            "type": "boolean"
          },
          "resolution_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolution Reason"
          },
          "status": {
            "anyOf": [
              {
                "enum": [
                  "open",
                  "acknowledged",
                  "escalated",
                  "closed"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "user": {
            "default": "analyst_1",
            "title": "User",
            "type": "string"
          }
        },
        "title": "IncidentPatchRequest",
        "type": "object"
      },
...
      "IncidentSubjectNew": {
        "properties": {
          "host": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Host"
          },
          "source_ip": {
            "title": "Source Ip",
            "type": "string"
          },
          "username": {
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "source_ip",
          "username"
        ],
        "title": "IncidentSubjectNew",
        "type": "object"
      },
...
      "InvestigationPivotQuery": {
        "properties": {
          "description": {
            "title": "Description",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "query": {
            "title": "Query",
            "type": "string"
          }
        },
        "required": [
          "name",
          "description",
          "query"
        ],
        "title": "InvestigationPivotQuery",
        "type": "object"
      },
      "LifecycleIncidentResponse": {
        "properties": {
          "affected_entities": {
            "items": {
              "type": "string"
            },
            "title": "Affected Entities",
            "type": "array"
          },
          "ai_pipeline_data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AIPipelineMeta"
              },
              {
                "type": "null"
              }
            ]
          },
          "assignee": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assignee"
          },
          "chain_confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chain Confidence"
          },
          "confidence": {
            "title": "Confidence",
            "type": "number"
          },
          "correlation_flags": {
            "items": {
              "type": "string"
            },
            "title": "Correlation Flags",
            "type": "array"
          },
          "created_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "dest": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dest"
          },
          "detection_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detection Id"
          },
          "effective_priority": {
            "title": "Effective Priority",
            "type": "string"
          },
          "evidence": {
            "$ref": "#/components/schemas/IncidentEvidenceNew"
          },
          "evidence_count": {
            "title": "Evidence Count",
            "type": "integer"
          },
          "explanation": {
            "$ref": "#/components/schemas/IncidentExplanationNew"
          },
          "first_seen": {
            "title": "First Seen",
            "type": "string"
          },
          "incident_id": {
            "title": "Incident Id",
            "type": "string"
          },
          "investigation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/IncidentInvestigationNew"
              },
              {
                "type": "null"
              }
            ]
          },
          "is_stale": {
            "title": "Is Stale",
            "type": "boolean"
          },
          "last_seen": {
            "title": "Last Seen",
            "type": "string"
          },
          "mitre": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/IncidentMitreMappingNew"
              },
              {
                "type": "null"
              }
            ]
          },
          "mitre_technique": {
            "default": "T1110",
            "title": "Mitre Technique",
            "type": "string"
          },
          "playbook_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Playbook Id"
          },
          "recommended_actions": {
            "items": {
              "type": "string"
            },
            "title": "Recommended Actions",
            "type": "array"
          },
          "resolution_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolution Reason"
          },
          "risk_level": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Risk Level"
          },
          "risk_score": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Risk Score"
          },
          "severity": {
            "title": "Severity",
            "type": "string"
          },
          "sla_action_required": {
            "title": "Sla Action Required",
            "type": "boolean"
          },
          "sla_breached": {
            "title": "Sla Breached",
            "type": "boolean"
          },
          "sla_deadline_minutes": {
            "title": "Sla Deadline Minutes",
            "type": "integer"
          },
          "source": {
            "default": "auth",
            "enum": [
              "auth",
              "ai_pipeline"
            ],
            "title": "Source",
            "type": "string"
          },
          "source_count": {
            "title": "Source Count",
            "type": "integer"
          },
          "status": {
            "default": "open",
            "enum": [
              "open",
              "acknowledged",
              "escalated",
              "closed"
            ],
            "title": "Status",
            "type": "string"
          },
          "subject": {
            "$ref": "#/components/schemas/IncidentSubjectNew"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          },
          "time_in_state_seconds": {
            "title": "Time In State Seconds",
            "type": "integer"
          },
          "time_since_alert_seconds": {
            "title": "Time Since Alert Seconds",
            "type": "integer"
          },
          "type": {
            "title": "Type",
            "type": "string"
          },
          "updated_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          },
          "valid_next_statuses": {
            "items": {
              "type": "string"
            },
            "title": "Valid Next Statuses",
            "type": "array"
          }
        },
        "required": [
          "incident_id",
          "type",
          "severity",
          "confidence",
          "first_seen",
          "last_seen",
          "affected_entities",
          "evidence_count",
          "source_count",
          "summary",
          "recommended_actions",
          "explanation",
          "subject",
          "evidence",
          "is_stale",
          "time_since_alert_seconds",
          "time_in_state_seconds",
          "sla_deadline_minutes",
          "sla_breached",
          "sla_action_required",
          "effective_priority",
          "valid_next_statuses"
        ],
        "title": "LifecycleIncidentResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "SecureWatch Engine",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/incidents/{incident_id}": {
      "get": {
        "operationId": "get_incident_incidents__incident_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "incident_id",
            "required": true,
            "schema": {
              "title": "Incident Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LifecycleIncidentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Incident",
        "tags": [
          "incidents"
        ]
      },
      "patch": {
        "operationId": "patch_incident_incidents__incident_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "incident_id",
            "required": true,
            "schema": {
              "title": "Incident Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IncidentPatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LifecycleIncidentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Patch Incident",
        "tags": [
          "incidents"
        ]
      }
    },
...
    "/validate/results/report": {
      "get": {
        "operationId": "download_report_validate_results_report_get",
        "parameters": [
          {
            "in": "query",
            "name": "org",
            "required": false,
            "schema": {
              "default": "Your Organization",
              "title": "Org",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Download Report",
        "tags": [
          "validation"
        ]
      }
    },
    "/validate/{scenario_name}": {
      "post": {
        "operationId": "validate_scenario_validate__scenario_name__post",
        "parameters": [
          {
            "in": "path",
            "name": "scenario_name",
            "required": true,
            "schema": {
              "title": "Scenario Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Validate Scenario",
        "tags": [
          "validation"
        ]
      }
    }
  }
}